% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OperaNumeros.R
\name{multiplica.dos.numeros}
\alias{multiplica.dos.numeros}
\title{multiplica.dos.numero}
\usage{
multiplica.dos.numeros(num1, num2)
}
\arguments{
\item{numeros}{sobre los que se desea obtener el resultado}
}
\value{
El resultado de la operacion de multiplicar los numeros elegidos
}
\description{
Funcion para realizar la multiplicacion de dos dos numeros
}
\examples{
multiplica.dos.numeros(4,6)
}

