% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OperaNumeros.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{suma.dos.numeros}
\usage{
suma.dos.numeros(num1, num2)
}
\arguments{
\item{numeros}{sobre los que se desea obtener el resultado}
}
\value{
El resultado de la operacion de sumar los numeros elegidos
}
\description{
Funcion para realizar la suma de dos numeros
}
\examples{
suma.dos.numeros(5,5)
}

